search: Add create_pills to search_pil.js.
The items for search pill contain 2 keys, diplay_value and search_string.
display_value is the description of the search string. This description
is obtained from the operator object by using Filter.describe.
create_item_from_search_string creates an item in the above format given
an operator. get_search_string_from_item does the reverse of that.

Adding all the operator information i.e the operator, operand and
negated fields along with the search_string and description was tried out.
It was dropped because it didn't provide any advantage as one had to
always calculate the search_string and the description from the operator.