search: Add search pill on selecting typeahead.
I tried out using the description from suggestions.lookup_table to
append a pill using appendValidatedData so that the description
had not to be calculated again. But the description in the suggestions
lookup contains html due to highlighting. This html is escaped when inputed
in a pill. An attempt was also made to remove the higlighting by replacing
the tags. But other espaced characters like &lt; also popped up, so it was
better to use appendValue (which calls Filter.describe)
NOTE: UI changes to search query box are pending. The search icon overlaps
with the first pill.