search: Refresh search pill in narrow.activate.
The pill population logic from haschange.js has been removed. Although
it worked for the cases when a reload occurs but it did not work in
cases where the narrow was triggered by something else. Clicking on the
stream list did not add any pills.
Pill population code is added to narrow.activate. Pills are not populated
if the narrow was triggered by search as search handles the addition and
removal of pill by itself. The reason for not handling search too in
narrow.activate is to avoid clearing the pills and repopulating them.
Example of some of the triggers for narrow.activate include
`restore draft`, `topic change`,`sidebar`.