help-beta: Add option to format mdx documents.
Partially fixes #35129.

Prettier could not be used because of prettier issue number 12209, not
linked here to avoid spam backlinks on the original issue. Prettier does
not have support for mdx v2 and v3.

The starting set of rules for this file were borrowed from
https://github.com/wooorm/remark-preset-wooorm, some of the rules
that we did not want were removed from the starting set. None of the
rules were giving an error on the starting set, but some rules that
were giving lots of warnings on the generated mdx were removed, we
can add them back later as and when required.

We have not inserted this in the main lint tool, we should do that in
the final cutover PR since we don't want the lint here to give any
unexpected warnings when people are linting stuff unrelated to the mdx
files.

This commit has been tested out on the current state of help center to
not produce any errors or warnings. The first run of format will produce
tons of warnings which are the issues being auto-fixed by the linter.
After that, the second run should produce zero errors.