help-beta: Seperate heading/description and ordered list into 2 files.
For included files, our conversion tool needs a file just to be an
ordered list to properly surround it with FlattenList. If there is
heading and description in the file, the conversion script will not
inser FlattenList. So, we just create two different include files in
this case, first for the heading/description and second for the ordered
list.