help: Move URL redirects from Django to Astro.
Fixes #35649.

We do not serve the astro build via Django and thus we have to move the
help center redirects. We get a warning by pagefind on all these
redirects that they will be ignored in the search results as they don't
have an HTML body. We can ignore this warning till the cutover and
create a followup issue to solve that warning if this commit goes
through main.

Astro redirects are just directories inside dist/ that have an
index.html doing the redirect. A redirect for night-mode to dark-mode
will mean that there is a dist/night-mode/index.html that has a redirect
with href="dark-mode". If we have a trailing slash in the URL, the
redirect would instead become help/night-mode/dark-mode instead of
redirecting help/night-mode to help/dark-mode. We remove the trailing
slash in our help center middleware.