help: Add production config for the new help center.
Most of the nginx config is copied from
https://docs.astro.build/en/recipes/docker/#nginx.

Astro redirects are just directories inside dist/ that have an
index.html doing the redirect. A redirect for night-mode to dark-mode
will mean that there is a dist/night-mode/index.html that has a redirect
with href="dark-mode". If we have a trailing slash in the URL, the
redirect would instead become help/night-mode/dark-mode instead of
redirecting help/night-mode to help/dark-mode. We take care of removing
the trailing slash in the nginx config.

@types/mdast was in dev dependencies while it shouldn't have been, we
move that away from dev dependencies and bump the version.

We ideally should be generating two builds for Zulip cloud and serve a
different one depending on whether the subdomain is same as the root
domain or not. We can look into this as an immediate followup while this
commit helps things get ready for a test deploy on CZO.