help: Add production config for the new help center.
Fixes #35131.

Most of the nginx config is copied from
https://docs.astro.build/en/recipes/docker/#nginx.

@types/mdast was in dev dependencies while it shouldn't have been, we
move that away from dev dependencies and bump the version.

We ideally should be generating two builds for Zulip cloud and serve a
different one depending on whether the subdomain is same as the root
domain or not. We can look into this as an immediate followup while this
commit helps things get ready for a test deploy on CZO.