help: Remove sidebar index from includes.
The sidebar index was an includes file in our previous help center since
we were using it to programmatically in documentation.py. Since that is
no longer the case, there is no reason for this file to be in includes.

A side-effect of us doing this is that now the overview section on the
right for the index page will be generated correctly with us avoiding
the issue in #35127.