help: Move static/images/help to starlight_help/src/images.
Astro was importing these images and doing it's own processing anyways,
there was not a reason for these images to stay in the static folder.

We duplicate channels-and-topics.png between starlight_help/src/images
and static/images/landing-page/why-zulip, this should be fine
considering another copy of the same image lies in
static/images/emails.